Deleted entity with uuid class
