--- src/main/java/com/auctionapp/api/model/entities/EntityWithUUID.java
@@ -1,33 +1,0 @@
-package com.auctionapp.api.model.entities;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-import org.hibernate.annotations.Type;
-
-@MappedSuperclass
-public abstract class EntityWithUUID implements Serializable {
-	@Id
-	@GeneratedValue(generator = "UUID")
-	@Type(type = "pg-uuid")
-	private UUID uuid;
-
-	public EntityWithUUID() {
-	}
-
-	public EntityWithUUID(UUID uuid) {
-		this.uuid = uuid;
-	}
-
-	public UUID getUuid() {
-		return uuid;
-	}
-
-	public void setUuid(UUID uuid) {
-		this.uuid = uuid;
-	}
-}
